# Conflicts:
#   tools/benchmark/src/main/java/fr/umlv/rootcause/benchmark/Benchmark.java

+++ core/src/main/java/fr/uge/modules/api/model/linking/Computation.java
@@ -1,4 +1,4 @@
 package fr.uge.modules.api.model.linking;
 
-public record Computation(String token_type, String value_log_first, String value_log_second, float proximity) {
+public record Computation(String token_type, String value_log_first, String value_log_second, Double proximity) {
 }
